# api/openapi.yaml
openapi: 3.0.3
info:
  title: Zalipuli API
  version: 0.1.0

servers:
  - url: http://localhost:8080/

paths:
  /health:
    get:
      operationId: getHealth
      summary: Healthcheck
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /levels/start:
    post:
      operationId: postLevelsStart
      summary: Generate a new level
      requestBody:
        content:
          application/json:
            schema:
              $ref:
                '#/components/schemas/StartLevelRequest'
      responses:
        '201':
          description: Level created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /levels/{levelId}:
    get:
      operationId: getLevel
      summary: Get level by id
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Level found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelResponse'
        '404':
          description: Level not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /levels/finish:
    post:
      operationId: finishLevel
      summary: Delete level by id
      requestBody:
        content:
          application/json:
            schema:
              $ref:
                '#/components/schemas/FinishLevelRequest'
      responses:
        '204':
          description: Level deleted
        '404':
          description: Level not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /levels/hint:
    post:
      operationId: postLevelsHint
      summary: Get hint for level
      requestBody:
        content:
          application/json:
            schema:
              $ref:
                '#/components/schemas/HintRequest'
      responses:
        '200':
          description: Hint found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HintResponse'


components:
  schemas:
    StartLevelRequest:
      type: object
      required:
        - gameName
      properties:
        gameName:
          $ref: '#/components/schemas/GameName'

    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok

    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string

    LevelResponse:
      type: object
      required:
        - id
        - status
        - startLevelState
      properties:
        id:
          type: string
          example: 43e8e3e6-a3d0-41f7-9516-9835565d7997
        status:
          type: string
          enum:
            - new
            - pending
            - ready
            - incorrect
        startLevelState:
          $ref: '#/components/schemas/LevelState'
        minSteps:
          type: integer
          minimum: 0

    GameName:
      type: string
      enum:
        - watersort

    Vials:
      type: array
      description: Array of vials, each vial is an array of colors from bottom to top
      items:
        type: array
        items:
          type: integer
          minimum: 1
          maximum: 15

    HintRequest:
        type: object
        required:
          - levelId
          - levelState
        properties:
          levelId:
            type: string
          levelState:
            $ref: '#/components/schemas/LevelState'


    LevelState:
      type: object
      oneOf:
        - $ref: '#/components/schemas/WaterSortLevelState'
      discriminator:
        propertyName: gameName
        mapping:
          watersort: '#/components/schemas/WaterSortLevelState'

    WaterSortLevelState:
      type: object
      required:
        - gameName
        - vials
      properties:
        gameName:
          $ref: '#/components/schemas/GameName'
        colorsCount:
          type: integer
          minimum: 5
          maximum: 15
        vials:
          $ref: '#/components/schemas/Vials'

    HintResponse:
      type: object
      required:
        - isSuccess
      properties:
        isSuccess:
          type: boolean
        hint:
          type: object
          oneOf:
            - $ref: '#/components/schemas/WaterSortHint'
          discriminator:
            propertyName: gameName
            mapping:
              watersort: '#/components/schemas/WaterSortHint'

    WaterSortHint:
      type: object
      required:
        - vialIndexFrom
        - vialIndexTo
        - gameName
      properties:
        gameName:
          $ref: '#/components/schemas/GameName'
        vialIndexFrom:
          type: integer
        vialIndexTo:
          type: integer

    FinishLevelRequest:
      type: object
      required:
        - levelId
      properties:
        levelId:
          type: string