# pkg/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

import (
	"encoding/json"
	"errors"

	"github.com/oapi-codegen/runtime"
)

// Defines values for GameName.
const (
	Watersort GameName = "watersort"
)

// Valid indicates whether the value is a known member of the GameName enum.
func (e GameName) Valid() bool {
	switch e {
	case Watersort:
		return true
	default:
		return false
	}
}

// Defines values for LevelResponseStatus.
const (
	Incorrect LevelResponseStatus = "incorrect"
	New       LevelResponseStatus = "new"
	Pending   LevelResponseStatus = "pending"
	Ready     LevelResponseStatus = "ready"
)

// Valid indicates whether the value is a known member of the LevelResponseStatus enum.
func (e LevelResponseStatus) Valid() bool {
	switch e {
	case Incorrect:
		return true
	case New:
		return true
	case Pending:
		return true
	case Ready:
		return true
	default:
		return false
	}
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Message string `json:"message"`
}

// FinishLevelRequest defines model for FinishLevelRequest.
type FinishLevelRequest struct {
	LevelId string `json:"levelId"`
}

// GameName defines model for GameName.
type GameName string

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status string `json:"status"`
}

// HintRequest defines model for HintRequest.
type HintRequest struct {
	LevelId    string     `json:"levelId"`
	LevelState LevelState `json:"levelState"`
}

// HintResponse defines model for HintResponse.
type HintResponse struct {
	Hint      *HintResponse_Hint `json:"hint,omitempty"`
	IsSuccess bool               `json:"isSuccess"`
}

// HintResponse_Hint defines model for HintResponse.Hint.
type HintResponse_Hint struct {
	union json.RawMessage
}

// LevelResponse defines model for LevelResponse.
type LevelResponse struct {
	Id              string              `json:"id"`
	MinSteps        *int                `json:"minSteps,omitempty"`
	StartLevelState LevelState          `json:"startLevelState"`
	Status          LevelResponseStatus `json:"status"`
}

// LevelResponseStatus defines model for LevelResponse.Status.
type LevelResponseStatus string

// LevelState defines model for LevelState.
type LevelState struct {
	union json.RawMessage
}

// StartLevelRequest defines model for StartLevelRequest.
type StartLevelRequest struct {
	GameName GameName `json:"gameName"`
}

// Vials Array of vials, each vial is an array of colors from bottom to top
type Vials = [][]int

// WaterSortHint defines model for WaterSortHint.
type WaterSortHint struct {
	GameName      GameName `json:"gameName"`
	VialIndexFrom int      `json:"vialIndexFrom"`
	VialIndexTo   int      `json:"vialIndexTo"`
}

// WaterSortLevelState defines model for WaterSortLevelState.
type WaterSortLevelState struct {
	ColorsCount *int     `json:"colorsCount,omitempty"`
	GameName    GameName `json:"gameName"`

	// Vials Array of vials, each vial is an array of colors from bottom to top
	Vials Vials `json:"vials"`
}

// FinishLevelJSONRequestBody defines body for FinishLevel for application/json ContentType.
type FinishLevelJSONRequestBody = FinishLevelRequest

// PostLevelsHintJSONRequestBody defines body for PostLevelsHint for application/json ContentType.
type PostLevelsHintJSONRequestBody = HintRequest

// PostLevelsStartJSONRequestBody defines body for PostLevelsStart for application/json ContentType.
type PostLevelsStartJSONRequestBody = StartLevelRequest

// AsWaterSortHint returns the union data inside the HintResponse_Hint as a WaterSortHint
func (t HintResponse_Hint) AsWaterSortHint() (WaterSortHint, error) {
	var body WaterSortHint
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromWaterSortHint overwrites any union data inside the HintResponse_Hint as the provided WaterSortHint
func (t *HintResponse_Hint) FromWaterSortHint(v WaterSortHint) error {
	v.GameName = "watersort"
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeWaterSortHint performs a merge with any union data inside the HintResponse_Hint, using the provided WaterSortHint
func (t *HintResponse_Hint) MergeWaterSortHint(v WaterSortHint) error {
	v.GameName = "watersort"
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

func (t HintResponse_Hint) Discriminator() (string, error) {
	var discriminator struct {
		Discriminator string `json:"gameName"`
	}
	err := json.Unmarshal(t.union, &discriminator)
	return discriminator.Discriminator, err
}

func (t HintResponse_Hint) ValueByDiscriminator() (interface{}, error) {
	discriminator, err := t.Discriminator()
	if err != nil {
		return nil, err
	}
	switch discriminator {
	case "watersort":
		return t.AsWaterSortHint()
	default:
		return nil, errors.New("unknown discriminator value: " + discriminator)
	}
}

func (t HintResponse_Hint) MarshalJSON() ([]byte, error) {
	b, err := t.union.MarshalJSON()
	return b, err
}

func (t *HintResponse_Hint) UnmarshalJSON(b []byte) error {
	err := t.union.UnmarshalJSON(b)
	return err
}

// AsWaterSortLevelState returns the union data inside the LevelState as a WaterSortLevelState
func (t LevelState) AsWaterSortLevelState() (WaterSortLevelState, error) {
	var body WaterSortLevelState
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromWaterSortLevelState overwrites any union data inside the LevelState as the provided WaterSortLevelState
func (t *LevelState) FromWaterSortLevelState(v WaterSortLevelState) error {
	v.GameName = "watersort"
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeWaterSortLevelState performs a merge with any union data inside the LevelState, using the provided WaterSortLevelState
func (t *LevelState) MergeWaterSortLevelState(v WaterSortLevelState) error {
	v.GameName = "watersort"
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

func (t LevelState) Discriminator() (string, error) {
	var discriminator struct {
		Discriminator string `json:"gameName"`
	}
	err := json.Unmarshal(t.union, &discriminator)
	return discriminator.Discriminator, err
}

func (t LevelState) ValueByDiscriminator() (interface{}, error) {
	discriminator, err := t.Discriminator()
	if err != nil {
		return nil, err
	}
	switch discriminator {
	case "watersort":
		return t.AsWaterSortLevelState()
	default:
		return nil, errors.New("unknown discriminator value: " + discriminator)
	}
}

func (t LevelState) MarshalJSON() ([]byte, error) {
	b, err := t.union.MarshalJSON()
	return b, err
}

func (t *LevelState) UnmarshalJSON(b []byte) error {
	err := t.union.UnmarshalJSON(b)
	return err
}